sudo check

+++ uninstall.sh
@@ -1,6 +1,13 @@
 #!/bin/bash
 
+if [ "$EUID" -ne 0 ]
+  then echo "Please run as sudo."
+  exit
+fi
+
 echo "Uninstalling WolkGatewayModule-Modbus..."
+service modbus_module stop
+echo "Stopped running service"
 rm /etc/systemd/system/modbus_module.service
 echo "Removed service unit file"
 systemctl daemon-reload
@@ -8,6 +15,6 @@ echo "Systemctl reloaded"
 rm -rf /etc/modbusModule/
 echo "Removed configuration files"
 rm -rf /usr/local/bin/lib/
-rm -rf /usr/local/bin/modbusModule
+rm /usr/local/bin/modbusModule
 echo "Removed application files"
 echo "Done."
